Cleanup to test for inserting ref temps
[proposed by hilde; implemented and tested by me]

This is a cleanup to r22213 suggested by Tom in his review which is a
slightly nicer way to test whether or not we need a ref temp.  Instead
of checking for the absence of a reference type in the actual, we
check if the actual's type is the value type for the formal's ref
type.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22214 3a8e244f-b0f2-452b-bcba-4c88e055c3ca