Refactor new_RealSymbol and new_ImagSymbol into new_FloatSymbol
[reviewed by hilde]

In reviewing my change to new_ImagSymbol in r22202, Tom pointed
out the similarities betweeen new_RealSymbol and new_ImagSymbol,
so as a post-commit cleanup, I refactored these to share a
helper routine, new_FloatSymbol to remove the redundancy.  As
a result of this refactoring the bug fix for imaginary literals
is now applied to real literals, though it should never actually
fire since real literals should always contain a decimal point
or e/E character (at least in the current language definition).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22215 3a8e244f-b0f2-452b-bcba-4c88e055c3ca