Remove dead locale memory interface
[reviewed in principle by hilde, gbt]

In a discussion about untracked memory allocations yesterday, the four
memory interface routines on the locale class came up as an instance
of untracked memory allocations; but it turns out that they are
currently unused (and perhaps never were -- at least, I can't find any
signs that they could be called from the compiler or ever were).  So
let's remove them to avoid fear that they lead to untracked
allocations, and to clean up.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22216 3a8e244f-b0f2-452b-bcba-4c88e055c3ca