Clean up locale initialization and fix race condition in multilocale startup
[reviewed by David and others]

This commit replaces the forall loop for locales initialization with a
coforall+on optimized loop.  It also fixes the race condition added in
r22171.

Using a coforall+on optimized loop reduces the number of
simultaneously created local tasks created during startup, since the
tasks are created only on the remote end.

The fix for the race condition required a barrier across the tasks
creating the locales objects (one per locale).

NOTES:

It would have been nice to be able to use the runtime function
chpl_comm_barrier(), but it can only handle one barrier at a time, and
the non-0 locales are already waiting in the exit barrier.  Instead, I
implemented a very specialized barrier in Chapel for this case.  The
thing that I would most like to improve upon is the broadcast when the
barrier is finished.  Right now, I'm using a for loop (because we
can't use parallel iteration at this point).

A few other weirdnesses due to bugs are commented in the file:

- Can't resolve constructors for nested records/classes
- Can't use a record with an array member due to initCopy needing to
  ChapelDistribution
- Bug with nested procedures and extern consts

During the review with the group, it occurred to me that I could try
to reuse the Locales array for this barrier.  I did a quick test of
one solution and while it was possible, it was not clearly a win
especially since it required more simultaneous tasks.  I'm going to
continue investigating that option as a background task, as I think it
will be cleaner in most ways.

TEST CHANGES:

- Removed config const minRootLocaleInitPerTask from -h output
- Update of comm counts during initialization.  We have now have a few
  more gets/puts/forks from remote locales for the barrier.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22232 3a8e244f-b0f2-452b-bcba-4c88e055c3ca