Replace flag-by-flag copy with a bulk flag copy in copyInner() for subclasses of Symbol. [reviewed by bradc]
The length of the list of flags on function symbols we want to copy when the
function is instantiated has grown to five, that on TypeSymbol is three.
At this point, it appears more efficient to bulk-copy all flags and then remove
them explicitly from the copy (in the client code) if needed.

This patch started out just doing the bulk copy for FnSymbol.  At Brad's
suggestion, I added the other subclasses of symbol (for which copyInner is
implemented).

[FnSymbol::copyInner: tested on test/ (standard configuration) 
 all implementations of Symbol::copyInner: tested on primers/ (standard configuration),
                                           tested on multilocale/ (gasnet).]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22235 3a8e244f-b0f2-452b-bcba-4c88e055c3ca