Fixed users/christopher.doris/intersperse.future.
[reviewed by Sung]

BACKGROUND

The above bug appeared when an iterator was not inlined, but instead
an iterator class and methods such as advance() got created.

r21676 introduced a couple of changes in the above case, among others:

(a) A PRIM_YIELD could contain a SymExpr with any symbol, including
globals (including literals and other params). Different PRIM_YIELDs
in the same iterator could reference different Symbols.

(b) All iterator-local Symbols were represented by the 'value' field
of the iterator class (aka ic.value or valField in iterator.cpp)

(c) Upon a yield, ic.value got updated when yielding a global symbol,
but not when yielding a local symbol. The assumption was that, since
updates to a local symbol were already reflected in ic.value,
no additional updating was needed.

The above resulted in incorrect settings of ic.value for example
in the following scenarios:

  loc1 = expr
  yield literal
  yield loc1

  loc1 = expr1
  loc2 = expr2
  yield loc1
  yield loc2

THE FIX

This commit leaves (a) unchanged, but changes (b) and (c).

The case when all yields in the iterator reference the same local symbol
is an exception - it is handled as before. The variables 'oneLocalYS'
are set in this case.

From the new comment in iterator.cpp:

oneLocalYS is set when all PRIM_YIELDs in the iterator
yield the same Symbol *and* that Symbol is defined locally
within the iterator.

When oneLocalYS: that Symbol is represented directly by ic.value.
I.e. the Symbol is mapped directly to valField, in
local2field, replaceLocalsWithFields, etc.
ic.value is not updated explicitly upon a yield - because it always
has the up-to-date value of that Symbol.

When !oneLocalYS: each local Symbol is represented by its own field
ic.FNN_xxx. Non-local Symbols are not represented in ic's fields.
Upon each yield, ic.value aka valField is explicitly updated
with yield's expression.
E.g. 'yield localvar' is converted to ic.value = ic.FNN_localvar.

Addenda:

If there are no yields in the iterator, it does not seem to matter
whether 'oneLocalYS' is set or not.

oneLocalYS is set in collectYieldSymbols().

I modified addLocalsToClassAndRecord() to implement the above logic.

WHILE THERE

There was a question in replaceLocalWithFieldTemp() why we need
to check for the case call->isPrimitive(PRIM_SET_MEMBER) explicitly
instead of relying on 'is_def'.

The answer seems that currently buildDefUseSets() does not identify
PRIM_SET_MEMBER and PRIM_SVEC_SET_MEMBER as defs, so we need to
compensate for that. I added the case for SVEC and cleaned up comments.

I did not address the buildDefUseSets issue because some users of
the information (e.g. dead code elimination) rely on all defs being
PRIM_MOVE and so do not handle PRIM_SET_MEMBER and PRIM_SVEC_SET_MEMBER.

TESTS

Un-futurized christopher.doris/intersperse.chpl, obviously.

Added a bunch of tests in functions/iterators/vass/multi-yield-symbols.
Many are obtained by copying and small tweaking.

  11-oneLocalSymbol      - all yields use the same symbol, which is local
  12-oneLocalOneGlobal   - within an iterator, there are yield(s) using
                           a local symbol, there are yield(s) using
                           a global symbol
  13-oneLocalOneLiteral  - same as 12, but using a literal instead of
                           the global symbol
  14-diffLocalSymbols    - within an iterator, yields use different
                           local symbols
  15-diffLocalsOneGlobal - a blend of 12 and 14
  16-diffLocalsOneLitera - same as 15, but using a literal instead of
                           the global symbol

(A "local symbol" is either a formal or a local variable.)

Future work/todo: add copies of these tests where helper() is replaced
with a range or a domain, which is either a formal (as in Chris's
original test) or a global.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22237 3a8e244f-b0f2-452b-bcba-4c88e055c3ca