Set up nightly numa testing
[reviewed by Elliot]

This patch sets up nightly testing on chap05 to include numa
localeModel examples.

In nightly, I added the -localeModel flag that says to use the locale
Mode suite (similar to the multilocale suite).  I also added the value
of CHPL_LOCALE_MODEL to the file names if it was not "flat".

The nightly.numa script runs the numa tests.  For now put in an exit
before running the gasnet tests, because when I tested it, they took
another 3 hours.

While I was in there, I changed the naming of the local tmp files in
nightly.fast, nightly. qthreads, and nightly.numa.  I put the username
and pid in the /tmp logfile name to avoid possible conflict.  The way
is it now, when someone runs the script for debugging purposes under
their own user id, you get permission denied due to the files in /tmp.
I also removed the tmp files if the copy to the central location was
successful.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22240 3a8e244f-b0f2-452b-bcba-4c88e055c3ca