Deprecate chpl_comm_alloc_registry() and simplify chpl_globals_registry
[reviewed by gbt and mppf]

In reviewing the communication interface recently, we noticed that all
implementations of chpl_comm_alloc_registry() are now identical,
suggeting that it's no longer a useful part of the communication
interface.  Looking at the archives, it appears that its original
intention was to permit the globals registry to be heap allocated, but
for whatever reason, we've gone away from that style of approach (I
believe it's because we've come up with more robust ways of dealing
with environments in which C isn't guaranteed to place globals at the
same base addresses across SPMD executions, as we saw a year or two
ago on the Mac).

This commit removes chpl_comm_alloc_registry(), and folds the
chpl_globals_registry and chpl_globals_registry_static variables into
a single variable (named chpl_globals_registry but declared as
chpl_globals_registry_static was before).  It also removes a compiler
primitive that was used to generate the call to
chpl_comm_alloc_registry() but is no longer required.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22242 3a8e244f-b0f2-452b-bcba-4c88e055c3ca