Update syncsingle test
Work done after 1.7 to reduce the number of things being put on the
heap got rid of the unnecessary back and forth to locale 0 (described
in the original commit of this test).  Now locale 0 does not need to
do any gets, and so we can remove the prediff that handled this case.

While I was investigating I added a config param for verbose comm into
the test.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22243 3a8e244f-b0f2-452b-bcba-4c88e055c3ca