Use FLAG_COMPILER_GENERATED instead of FLAG_TEMP to mark compiler-generated function and label symbols. [reviewed by vass]
Prior to this patch, we overloaded the meaning of FLAG_TEMP: attached to
variable symbols, it means an unnamed temporary; attached to label or
function symbols, it mas the more basic meaning of an AST element added by the
compiler.

When used with function symbols, its main purpose is to distinguish
compiler-supplied functions from user-supplied versions.  During resolution,
this allow the compiler-supplied versions of some functions (like the
default-initializer function for records and classes) to be overridden by a
user-supplied version.

When applied to a variable symbol, it has a deeper meaning.  The rules for
user-managed memory allow the copy-initialization and destruction of an unnamed
temporary to be removed (or never inserted in the first place), because the
lifetime of such variables are known and their ownership never escapes the
function in which they are defined.

Since it is possible to envision a language construct where a variable is
user-generated but still (according to language rules) is treated as a
temporary, it is good to split these two uses.  Not to mention that it renders
the code more self-documenting. 


In splitting the two uses, I had to take care to mention both flags if the
existing code could apply to both {function and label symbols} and variable
symbols.  The only case where I found this to be true was in
addToSymbolTable().  

[tested with test/ (std)]


DETAILS
  
    compiler/include/flags_list.h
Added FLAG_COMPILER_GENERATED.

    compiler/AST/build.cpp
Replaced FLAG_TEMP with FLAG_COMPILER_GENERATED on lines 539, 543, 569, 572,
932 and 936.  This means that compiler-generated labels now use the "compiler
generated" flag in place of the "temp" flag.

    compiler/passes/normalize.cpp
Added a comment on how FLAG_TEMP is used.
 
    compiler/passes/insertLineNumbers.cpp
The determination whether to insert filename/line-number information in a call
is made w.r.t. whether the function being called is internal or user-defined.
Since it pertains to function symbols, FLAG_COMPILER_GENERATED is the right flag
to test in this context.
 
    compiler/passes/scopeResolve.cpp
In addToSymbolTable(), added the flag "compiler generated" to cover the possible
case that some of the DefExprs passed in refer to label or function symbols.
For readability, I changed
  if (!def->sym->hasFlag(FLAG_TEMP) &&
      !def->sym->hasFlag(FLAG_COMPILER_GENERATED)) { ... }
to
  if (def->sym->hasFlag(FLAG_TEMP) ||
      def->sym->hasFlag(FLAG_COMPILER_GENERATED))
    continue;
    ...
and adjusted the indentation of the surrounding code.

In build_type_constructor() and build_constructor(), I changed FLAG_TEMP to
FLAG_COMILER_GENERATED, since these are both functions.

    compiler/passes/checkParsed.cpp
Added a comment about the use of FLAG_TEMP.

    compiler/passes/buildDefaultFunctions.cpp
Replaced FLAG_TEMP with FLAG_COMPILER_GENERATED on lines 160, 313, 338, 477, and
522.  These are all compiler-generated functions.  Added the "compiler
generated" flag to several other functions generated in this pass.

Ideally, all of the functions generated in this pass would get the "compiler
generated" flag.  As it stands, there is some contention with default versions
defined in the module code.  The desired compiler behavior is to choose the
compiler-generated version over the default version supplied in the module
code.  But resolution says the module version is preferred.

These places were noted in the code with TODO comments.  It is expected that in
a future patch the module default version will be removed or marked with
"compiler generated".  After that, it will be possible to mark all the functions
generated here with the "compiler generated" flag, and the TODO comments removed.

    compiler/resolution/wrappers.cpp
Replaced FLAG_TEMP with FLAG_COMPILER_GENERATED on lines 86 and 155.  These both
apply to wrapper functions.

    compiler/resolution/functionResolution.cpp
Replaced FLAG_TEMP with FLAG_COMPILER_GENERATED on lines 1820, 1823, 1975, 2014,
2331 and 2351.  All of these tests relate to function symbols.

    compiler/util/misc.cpp
Replaced FLAG_TEMP with FLAG_COMPILER_GENERATED on lines 151 and 206.  These
tests both relate to function symbols.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22246 3a8e244f-b0f2-452b-bcba-4c88e055c3ca