Future test cleanup
Vass pointed out that this test was buggy (didn't capture what it was
meant to) and that the .good was missing a newline and asked me to fix
those issues which I did.  While here, I also updated the .future
message to reflect the last 5 years of thought about the language and
added a .bad file.

I also changed it from an "error message" future to a "bug" future
because the former are typically used when a future generates an error
message but that message should be clarified, whereas a case like this
where the test should generate an error message but doesn't is more a
case of buggy behavior.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22248 3a8e244f-b0f2-452b-bcba-4c88e055c3ca