Improve sharedRefCount2 test. [suggested by vass]
The previous version of the test relied on freed memory being reused immediately
and then overwritten.  It's up to the allocator whether freed memory is used
again right away, so the test did not fail reliably.

This was remedied by adding a flag to the implementation, to indicate if an
implementation is accessible after it is freed.

As written the test will still fail to fail (i.e. falsely report success) if the
allocator overwrites freed memory with zeroes.  This for example can be done in
the valgrind allocator by passing the --free-fill=00 option.   (Using anything
other than zero for the fill pattern should still elicit the error.)


DETAILS

    sharedRefCount2.chpl
Just test the freed flag in the implementation pointed by the handle returned
from escape().  If so, report an error.  (We should not be accessing freed
memory.)  If not, report success.

    refCountedType.chpl
Added the "freed" flag to the ref counted type implementation.  This flag is set
to true if the reference count has gone to zero, right before the implementation
is freed.  The helper routine release_helper() is used to do this consistently
between the handle destructor and the handle copy-assignment operator.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22250 3a8e244f-b0f2-452b-bcba-4c88e055c3ca