Repair runtime initiation of memory tracking.
[Reviewed by Sung.]

In r22223 I moved the initiation of memory tracking from the module code
to the runtime.  However, that broke a few tests, irretrievably in one
case: distributions/robust/arithmetic/reindexing/simple-alias-mem-use.
That is apparently the only test where we turn on memory tracking by
giving -smemTrack at compile time, and the emitted code for the
MemTracking module defaults its memTrack config const to true in the
absence of any explicit setting at execution time.  This wasn't handled
properly by the new code, because it was only looking at the config vars
known to the runtime.  In the simple-alias-mem-use case the memTrack
config var is not specified (either true or false) at execution time, so
the runtime doesn't have a set value for it.  As a result we didn't turn
on memory tracking, and the output for the test was incorrect.

Here we fix this problem, by once again making the runtime rely on the
values of the module's copies of the memtracking config consts, rather
than on the runtime's limited notion of the config const values.

Note: We should move the CHPL_TASK_STD_MODULES_INITIALIZED() call from
      where it currently is in chpl_main() to the pre-user-code hook.  I
      didn't mention this when I added that hook, but it does seem like
      the natural place for that call to occur.  (In fact, I can imagine
      a generalized pre-user-code callback capability, for the various
      parts of the runtime to register functions to be called right
      before the user code starts.)

The specific file changes are as follows.

modules/internal/MemTracking.chpl
  Add chpl_memTracking_returnConfigVals(), an exported proc the runtime
  can call when it wants the values of the memtracking config consts.
  The values are returned via ref arguments.  As an associated change,
  make memMax and memThreshold int(64), so that we can explicitly match
  the types for these in the caller and callee.

runtime/src/main.c
  Add a barrier at the beginning of the pre-user-code hook function, so
  that on non-0 locales the runtime memtracking code doesn't try to pick
  up the config consts from the module before module initialization has
  occurred and those values are set.

runtime/include/chplmemtrack.h
runtime/src/chplmemtrack.c
  Instead of getting the memTracking config consts from the runtime,
  call chpl_memTracking_returnConfigVals() to get them from the module
  code.

  While here, I also moved chpl_setMemFlags() to the beginning of
  chplmemtrack.c, where it belongs as an initialization function.

  Also while here, I tidied up some loose type handling.  I made memMax
  and memThreshold be size_t, to match the other variables they interact
  with.  And, I made a clean distinction between what are chpl_string
  variables and what are c_string ones, and converted as necessary from
  the former to the latter.  Right now those two types are the same, but
  someday they might not be and hopefully the new code will work if and
  when that happens.

  I also cleaned up some really long lines, because I had a hard time
  reading them.

test/memory/figueroa/LeakedMemory6.skipif
  Don't test this unless comm=none, because it throws --no-inline and
  with comm!=none that causes us to run afoul of an unrelated problem
  involving coerce and promotion wrappers (for which we already have a
  separate future).

test/memory/shannon/printFinalMemStat.no-local.good
  Fix one more test result that I missed in r22223.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22257 3a8e244f-b0f2-452b-bcba-4c88e055c3ca