Emergency fix for r22257 runtime build problem.
[Not reviewed.]

In a call in code committed in r22257 I'm passing the addresses of two
size_t (which is an unsigned int(64) type) arguments to a function that
is expecting signed int(64) pointers.  With CHPL_DEVELOPER set, this
fails compilation due to the mismatch.  Fix this for now by going
through signed int(64) intermediate args.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22260 3a8e244f-b0f2-452b-bcba-4c88e055c3ca