Fix the Plain-Old-Data optimization. [reviewed by ]
This fixes the failing test case types/records/hilde/sharedRefCount2.chpl, and
also gets past a crash that was occurring in users/ferguson/refcnt.chpl.

  Explanation:

There was an optimization starting on line 282 of
removeUnnecessaryAutoCopyCalls.cpp introduced by the comment
//
// remove pointless initCopy calls, e.g., initCopy calls on records of
// primitive types
//
whose intention was to avoid calling chpl__initCopy when the expectation
was that it would be trivial.  Triviality of chpl__initCopy was tested by
determining if the type of the argument was a fundamental type or a record
whose fields were of fundamental type (recursively).

This worked fine when all definitions of chpl__initCopy intersected by that test
were trivial.  However, we now have module code that meets the test and yet is
nontrivial.  Also, from a language design standpoint, if the user supplies a
nontrivial chpl__initCopy routine we ought not to skip it.

I chose the name for this optimization based on a discussion in the original C++ spec.
regarding when the elision of initialization (default- or copy-construction) is
legal.  The pertinent section is 6.7 [stmt.dcl], which treats the conditions
under which a goto can cross the initialization of a local variable.

The condition placed on such a crossing is that the associated type be POD and
have neither a nontrivial assignment operator nor a nontrivial destructor.  The
reason for this is that the destructor and assignment operator both assume that
an object has been initialized to a known state.  If the initialization code is
not run, that assumption may be violated.  

For our purposes, we should also not remove an initCopy call if chpl__initCopy()
for that type is nontrivial.  The reason is that, presumably, if the user takes
the trouble to define a copy-constructor, he expects to have it executed.  The
same reasoning will apply to user-defined default constructors, when they are
fully implemented.

Since it could be difficult or at least time consuming to determine if the
implementations of these three functions are all trivial, we use a weaker test:
If any is user-defined, then we cannot skip the initCopy call.  In this way, we
err on the side of caution.

Further introspection suggests that for one of these functions to be trivial,
its behavior must exactly match that of the compiler-supplied default version.
The test for this could still be expensive, so my recommendation is that we
state in the spec that the user's version is treated as nontrivial even if it
does match the compiler-supplied version's behavior exactly.  A test that is
simple for the compiler to apply is also simple for the user to apply.  So, 
"Why didn't the compiler apply the usual POD optimization?" should lead to "Oh,
I defined one of the four special functions that makes a type non-POD."


The application of "compiler generated" to the default implementation of
assignment on line 244 of ChapelBase.chpl resulted in an improvement in the
output of studies/madness/aniruddha/madchap/recordBug/test_likepy.  The
error-reporting code determines how far up the call stack to go before
attributing the whole-array assignment warning message.  Previously, the default
implementation of proc =() was considered to be a user-level function.  So in a
class or record with a field of array type but without an explicit assignment
operator, the warning message would be attributed to the assignment operator
expanded for that class.  Now that the "compiler generated" flag is applied to
the default implementation of proc =(), attribution of the whole-array
assignment warning propagates up through the array assignment to the
initialize() function that calls it (in this test case).


  Implementation:

The implementation proceeds in two steps: 1) First, we mark class (including
record and union) types that contain user-defined assignment and/or destructor
functions; 2) then, we scan for and remove initCopy and autoCopy calls whose
arguments are instances of those classes.  

In determining if an assignment operator affects a given type we examine only
the type of the LHS argument.  The RHS argument is assumed to be read-only.

In this context, "removal" means that AST construct like
  ('move' dst ("chpl__initCopy" src))
  ('move' dst ("chpl__autoCopy" src))
are replaced by
  ('move' dst src)
  ('move' dst src)
respectively.  This is basically a bit-wise copy of the argument into the
destination.  In UMM terms, ownership of the bits patter is shared between the
two variables.  When the copy (dst) goes out of scope, it does not have to be
destroyed.

The existing implementation made no attempt to remove the matching autoDestroy
when the initCopy call was skipped.  This new implementation adds that by
removing from affected variables the flags that cause autoDestroy calls to be
added.

To implement this, two new flags were added: FLAG_HAS_ASSIGNMENT and
FLAG_HAS_DESTRUCTOR.  FLAG_HAS_ASSIGNMENT is added to a ClassType that defines a
nontrivial assignment operator (as determined by the type of the LHS argument).
FLAG_HAS_DESTRUCTOR is added if the class (or record or union) defines a
nontrivial destructor.  


  Details:

  test/studies/madness/aniruddha/madchap/recordBug/test_likepy.good
The "whole-array assignment" warning message is now attributed to the line
containing the whole-array assignment in FTree.initialize(), due to propagation
(in FnSymbol::copyInner()) of the "compiler generated" flag.

  test/types/records/hilde/sharedRefCount2.future
  test/types/records/hilde/sharedRefCount2.bad
Removed.  This test now succeeds.

  modules/internal/ChapelBase.chpl
Added pragma "compiler generated" to the default implementation of proc =().
This makes the optimization treat it as trivial (which it is).

  compiler/include/flags_list.h
Added FLAG_HAS_ASSIGNMENT -- when attached to a class symbol, indicates that the
class defines a nontrivial assignment operator.
Added FLAG_HAS_DESTRUCTOR -- when attached to a class symbol, indicates that the
class defines a nontrivial destructor.
Corrected the comment on FLAG_AUTO_COPY_FN, which said "auto destroy function".

  compiler/passes/normalize.cpp
Slight rewrite, to give a name (ct) to thisDef->sym->type.

  compiler/passes/buildDefaultFunctions.cpp
Added the "compiler generated" flag to the compiler-generated record copy
function.

  compiler/optimizations/removeUnnecessaryAutoCopyCalls.cpp
Added function markPODtypes() to mark ClassTypes containing nontrivial
assignment operators and destructors with corresponding flags.
Revised isPrimitiveCopy() to return true only if the named type is a numerical
or class type or a record type that defines neither a nontrivial assignment
operator nor a nontrivial destructor and that contains only fields of numerical
or class type, or a similar record type (recursively).

Pulled out the removal of trivial initCopy calls as a separate routine, now called
removePODinitDestroy().  This uses the code from the rewritten version that
removes trivial autoCopy calls as well as trivial initCopy calls.
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22268 3a8e244f-b0f2-452b-bcba-4c88e055c3ca