Do memory reporting after runtime tasking layer shutdown, instead of before.
[Reviewed by Tom.]

This moves the memory reporting (leaks, etc.) call from right before we
shut down the tasking layer to right after.  The result is that we will
no longer appear to leak all the memory allocations done by the tasking
layer, at least to the extent it releases memory it allocates.

The only output change caused by this commit is in the leaked memory
report for the future memory/vass/memleak-array-of-records-1.  It still
produces the wrong output, but it's different wrong output in that the
leaked memory related to tasking is no longer there.  This is a plus,
because the point of the future is actually leaked memory related to
domains, which is still present and is more obvious in the absence of
the tasking related leakage.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22269 3a8e244f-b0f2-452b-bcba-4c88e055c3ca