Added a bug future for casting a nil class to a subclass which causes a segfault
From .future:
When the argument of a cast is nil, no error checking is performed to see if
the arguement is nil, even without --fast.

We are currently unsure if this should be a runtime error or if the cast should
return nil. If the argument is not nil, but the cast does not succed, should
that be an error, or should that return nil.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22276 3a8e244f-b0f2-452b-bcba-4c88e055c3ca