Fix for the performance regressions introduced by r22268. [reviewed by vass]
Revision 22268 makes a certain optimization for removing unnecessary initCopy
and autoCopy calls less aggressive.  The appearance of more initCopy (and
autoCopy) calls in the generated code is causing the observed increase in
communication.

In the revision, the optimization is applied only if the type is considered
Plain-Old-Data (POD). A record is considered POD (in this revision) if it
explicitly defines neither a destructor nor an assignment operator.  The flag
"compiler generated" can also be applied artificially, to inform the compiler
that an explicit implementation for an assignment operator or destructor is
trivial (for the purposes of this optimization).

Analysis of the increased comm counts showed that the additional initCopy call
(there was only one per failing test) had variable of range type as its
operand.  The solution was to apply the "compiler generated" flag to the
assignment operators defined for the range and rangeBase types.

[tested with 
 CHPL_COMM=gasnet in test/performance/ferguson/
 standard configuration in test/
]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22283 3a8e244f-b0f2-452b-bcba-4c88e055c3ca