Adds a feature request future for changing the single element array syntax to [x,]
From .future:
The current syntax for single element array literals can be a source of
confusion. For example when used with a for loop the thought can be to write
    for i in [1..2] do
        writeln(i);
which will return
    1..2
rather than the expected
    1
    2

By changing the syntax to be [x,] it becomes clear as to what is happening.
This also brings arrays into line with single element tuples - making things
more consistent across the language. In addition, there may be gains in
simplifying the parser with respect to the shorthand form of forall loops.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22286 3a8e244f-b0f2-452b-bcba-4c88e055c3ca