Add a feature request and semantic futures on specifiying types for the index value of an interation
From .futures:
Currently if you want the index variable to be a different size you need to cast an
element of the range to be of that type, e.g:
	for i in 1..10:int(8) do
		foo(i);
This is slightly unintuitive, allowing the user to specify the type via
	for i:int(8) in 1..10 do
		foo(i);
demonstrates a clearer intent as to what the user is trying to achieve.

However, when iterating over an array the loop index is treated as reference
into the array allowing for code such as
	var A = [1, 2, 3, 4];
	for i in A do
		i += 1;
to increment each element. When giving the index variable a type, such as
	for i:int(8) in A do
		i += 1;
a number of problems arise. For example if you were to do i += 400, what is
the result inside of the array A? So one option would be to treat any index
var given a type as being const. Another would be that the index type is const,
but only if it does not match the element type in the array. (In the above
example i:int(64) would not be const while i:int(8) would be)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22291 3a8e244f-b0f2-452b-bcba-4c88e055c3ca