Add initCopy functions to POD optimization test.
In deciding whether an initCopy call can be elided as trivial, we should also
take into account whether the initCopy function itself is defined in user code.

The argument for adding the initCopy function to this set is, if the user
defines nontrivial behavior for the initCopy function, then we had better call
it where initCopy behavior is required by the language.  

Updating the specification to state precisely when the user can expect his
initCopy function to be called is left as a future task.  As a hint, the
compiler may insert any number of local temporary variables, and the
initialization and destruction of these temporaries should not be visible to the
user.

It should be noted that this optimization would be unnecessary if the
compiler-generated versions were truly trivial.  In that case, the inlined
version of the initCopy function would do nothing more than the "optimized-out"
version -- specifically, yielding a bit-wise copy of the the original.


DETAILS

    modules/internal/ChapelBase.chpl
Added the "compiler generated" pragma to the default chpl__initCopy() function.

    compiler/include/flags_list.h
Added a new FLAG_HAS_USER_INIT_COPY_FN, to mark class types that contain a
user-defined initCopy() function.

    compiler/AST/iterator.cpp
Removed code that redundantly adds void returns to the four zip# functions.
This change should have been part of r22133.
It is required by the addition of the normalize call in ProtoIteratorFns().

    compiler/passes/buildDefaultFunctions.cpp
The FLAG_COMPILER_GENERATED is now added to chpl__initCopy functions generated
internally.

    compiler/resolution/functionResolution.cpp
Iterator functions are now normalized.
This is required by the call to getReturnSymbol() in cullAutoDestroyFlags() and
cullExplicitAutoDestroyFlags(),

    compiler/resolution/callDestructors.cpp
Added new functions cullAutoDestroyFlags() and cullExplicitAutoDestroyFlags() to
remove the corresponding flags from return value variables or variables that are
copied into the return value.

insertAutoDestroyTemps() was renamed as insertAutoDestroyCalls().

callDestructors() was factored into smaller functions.  (No functional changes)

    compiler/optimizations/removeUnnecessaryAutoCopyCalls.cpp
An additional clause was added in markPODtypes() to handle initCopy functions.
The type of the operand of the initCopy function is marked with the
FLAG_HAS_USER_INIT_COPY.
This flag is tested in isPrimitiveCopy() to determine if the corresponding type
is Plain-Old Data (POD).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22292 3a8e244f-b0f2-452b-bcba-4c88e055c3ca