Enable -fno-strict-overflow flag only for GCC versions 4.2 and greater.
The -fstrict-overflow flag was introduced in GCC version 4.2.
It is on by default when optimizations (-O2) are enabled.
It allows the compiler to assume that integer sums will not overflow.
We use -fno-strict-overflow currently, because the index and endpoint
calculations in range traversal (rangeBase.these) is expected to overflow when
traversing a maximal range.  See r22290.

[tested with hello.chpl on darwin and linux64]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22293 3a8e244f-b0f2-452b-bcba-4c88e055c3ca