Adding two versions of spectral norm.
spectral-norm-double-time.* follows the precedent of gcc version 4 and performs
the computation twice per process instead of once per process.  This shows an
improvement on our smaller testing machines, so it seems useful to keep it
around even though the improvement is (as expected) not seen on our largest
testing machine.

spectral-norm-barrier.* does the same thing as double-time, except it is based
off of the barrier implementation instead of lydia/spectral-norm.chpl.

Both of these versions have been added to the spectral norm graph, and data
points have been relabelled to more accurately represent the differences between
each implementation.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22295 3a8e244f-b0f2-452b-bcba-4c88e055c3ca